Implement a flush interval "heartbeat" table to identify all flush interval times
Counters are only sent in a flush interval if a UDP message was received
in that interval. Flush interval durations vary based on how the program
is configured and cannot be assumed to be a fixed value.

This commit records all flush intervals to a table within the hot metrics DB
so that missing counters can be identified.